# node-ts-mysql-api/node-ts-mysql-api/node-ts-mysql-api/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0

servers:
  - url: http://localhost:4000
    description: Local development server

paths:
  /accounts/authenticate:
    post:
      summary: Authenticate account credentials and return a JWT token and a cookie with a refresh token
      description: Accounts must be verified before authenticating.
      operationId: authenticate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "jason@example.com"
                password:
                  type: string
                  example: "pass123"
              required:
                - email
                - password
      responses:
        "200":
          description: Account details, a JWT access token and a refresh token cookie
          headers:
            Set-Cookie:
              description: "`refreshToken`"
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  title: { type: string }
                  firstName: { type: string }
                  lastName: { type: string }
                  email: { type: string }
                  role: { type: string }
                  created: { type: string }
                  isVerified: { type: boolean }
                  jwtToken: { type: string }
        "400":
          description: The email or password is incorrect
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Email or password is incorrect" }

  /accounts/refresh-token:
    post:
      summary: Use a refresh token to generate a new JWT token and a new refresh token
      description: The refresh token is sent and returned via cookies.
      operationId: refreshToken
      parameters:
        - in: cookie
          name: refreshToken
          description: The `refreshToken` cookie
          schema:
            type: string
      responses:
        "200":
          description: Account details, a JWT access token and a new refresh token cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwtToken: { type: string }
        "400":
          description: The refresh token is invalid, revoked or expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Invalid token" }

  /accounts/revoke-token:
    post:
      summary: Revoke a refresh token
      description: Admin users can revoke the tokens of any account, regular users can only revoke their own tokens.
      operationId: revokeToken
      security:
        - bearerAuth: []
      parameters:
        - in: cookie
          name: refreshToken
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
      responses:
        "200":
          description: The refresh token was successfully revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Token revoked" }
        "400":
          description: The refresh token is invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Invalid token" }
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  /accounts/register:
    post:
      summary: Register a new user account and send a verification email
      description: The first account registered is assigned the `Admin` role, others get `User`.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, example: "Mr" }
                firstName: { type: string, example: "Jason" }
                lastName: { type: string, example: "Watmore" }
                email: { type: string, example: "jason@example.com" }
                password: { type: string, example: "pass123" }
                confirmPassword: { type: string, example: "pass123" }
                acceptTerms: { type: boolean }
              required:
                - title
                - firstName
                - lastName
                - email
                - password
                - confirmPassword
                - acceptTerms
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Registration successful, please check your email for verification instructions" }

  /accounts/verify-email:
    post:
      summary: Verify a new account with a verification token received by email after registration
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
              required:
                - token
      responses:
        "200":
          description: Verification successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Verification successful, you can now login" }
        "400":
          description: Verification failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Verification failed" }

  /accounts/forgot-password:
    post:
      summary: Submit email address to reset the password on an account
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, example: "jason@example.com" }
              required:
                - email
      responses:
        "200":
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Please check your email for password reset instructions" }

  /accounts/validate-reset-token:
    post:
      summary: Validate the reset password token received by email
      operationId: validateResetToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
              required:
                - token
      responses:
        "200":
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Token is valid" }
        "400":
          description: Token is invalid
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Invalid token" }

  /accounts/reset-password:
    post:
      summary: Reset the password for an account
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                password: { type: string, example: "newPass123" }
                confirmPassword: { type: string, example: "newPass123" }
              required:
                - token
                - password
                - confirmPassword
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Password reset successful, you can now login" }
        "400":
          description: Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Invalid token" }

  /accounts:
    get:
      summary: Get a list of all accounts
      description: Restricted to admin users.
      operationId: getAllAccounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: An array of all accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    post:
      summary: Create a new account
      description: Restricted to admin users.
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string }
                password: { type: string }
                confirmPassword: { type: string }
                role: { type: string, enum: [Admin, User] }
              required:
                - title
                - firstName
                - lastName
                - email
                - password
                - confirmPassword
                - role
      responses:
        "200":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Email already registered
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  /accounts/{id}:
    parameters:
      - in: path
        name: id
        description: Account id
        required: true
        schema:
          type: string
    get:
      summary: Get a single account by id
      description: Admin users can access any account, regular users are restricted to their own account.
      operationId: getAccountById
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    put:
      summary: Update an account
      description: Admin users can update any account including role, regular users are restricted to their own account and cannot update role.
      operationId: updateAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                firstName: { type: string }
                lastName: { type: string }
                email: { type: string }
                password: { type: string }
                confirmPassword: { type: string }
                role: { type: string, enum: [Admin, User] }
      responses:
        "200":
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      summary: Delete an account
      description: Admin users can delete any account, regular users are restricted to their own account.
      operationId: deleteAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Account deleted successfully" }
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Account:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        role: { type: string }
        created: { type: string }
        updated: { type: string }
        isVerified: { type: boolean }
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: "Unauthorized" }
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string, example: "Not Found" }